starter/requirements.txt
alembic==1.4.3
aniso8601==6.0.0
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
authlib==0.15.2
awscli==1.18.112
Babel==2.8.0
botocore==1.17.35
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==3.0
DateTime==4.3
distlib==0.3.0
Django==3.0.7
docutils==0.15.2
ecdsa==0.15
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.7
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.1.0
pytz==2019.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.5
s3transfer==0.3.3
six==1.12.0
simplejson==3.17.2
SQLAlchemy==1.3.4
sqlparse==0.3.1
urllib3==1.25.10
virtualenv==20.0.23
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
zope.interface==5.1.0